separate unit from deployment tests
it has no sense to execute full tests together. we start here separation between two type of tests. locally, by executing test with command "pytest -m unittest" or "pytest -m deploymenttest", the separation works. for the moment on github actions, it is expected only to execute unit test, as the workflow to deploy automatically is not yet set.